Fix bug in move selection.
Only the first move in the list was actually a maximum-value move, because of
the way alpha-beta pruning works. To get a full set of maximum-value moves, we
would have to call Search(..., -hi, -lo + 1), otherwise if Search(...) returns
-lo, we don't know if the actual value is less than lo, or equal to it.

This bug didn't manifest itself because currently SelectMove() always returned
the first element of best_moves, but returning any other element would not be
correct.

Since I want to get maximum cut-off opportunities from alpha-beta pruning,
I decided to keep the search logic as it is, and simply return the first
maximum-value move as the best move. To get some randomization, it may be
necessary to shuffle the order of the fields around (which can probably be
done as part of move ordering, to be implemented later).