Add --logs option, to redirect player debug output to a logfile.
The option takes a filename prefix argument, so that if --logs=/tmp/log-,
then the output for the first game will be written to
/tmp/log-0000_0_red and /tmp/log-0000_1_blue for the first and second
player, respectively. Here, 0000 is the game number, 0 and 1 are the
player command numbers, and red and blue are the colors played by each
player in the corresponding game.